Move some runtime services types to uefi-raw
These are all direct moves, no changes to the types.
