+++ uefi-raw/src/table/mod.rs
@@ -4,6 +4,7 @@ mod header;
 mod revision;
 
 pub mod boot;
+pub mod runtime;
 
 pub use header::Header;
 pub use revision::Revision;
