+++ uefi/src/table/runtime.rs
@@ -2,12 +2,14 @@
 
 use super::{Header, Revision};
 use crate::table::boot::MemoryDescriptor;
-use crate::{guid, CStr16, Char16, Error, Guid, Result, Status, StatusExt};
+use crate::{CStr16, Char16, Error, Guid, Result, Status, StatusExt};
 use bitflags::bitflags;
 use core::ffi::c_void;
 use core::fmt::{Debug, Formatter};
 use core::mem::MaybeUninit;
 use core::{fmt, ptr};
+
+pub use uefi_raw::table::runtime::{TimeCapabilities, VariableAttributes, VariableVendor};
 
 #[cfg(feature = "alloc")]
 use {
@@ -601,74 +603,6 @@ impl PartialEq for Time {
 }
 
 impl Eq for Time {}
-
-/// Real time clock capabilities
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-#[repr(C)]
-pub struct TimeCapabilities {
-    /// Reporting resolution of the clock in counts per second. 1 for a normal
-    /// PC-AT CMOS RTC device, which reports the time with 1-second resolution.
-    pub resolution: u32,
-
-    /// Timekeeping accuracy in units of 1e-6 parts per million.
-    pub accuracy: u32,
-
-    /// Whether a time set operation clears the device's time below the
-    /// "resolution" reporting level. False for normal PC-AT CMOS RTC devices.
-    pub sets_to_zero: bool,
-}
-
-bitflags! {
-    /// Flags describing the attributes of a variable.
-    #[repr(transparent)]
-    #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord)]
-    pub struct VariableAttributes: u32 {
-        /// Variable is maintained across a power cycle.
-        const NON_VOLATILE = 0x01;
-
-        /// Variable is accessible during the time that boot services are
-        /// accessible.
-        const BOOTSERVICE_ACCESS = 0x02;
-
-        /// Variable is accessible during the time that runtime services are
-        /// accessible.
-        const RUNTIME_ACCESS = 0x04;
-
-        /// Variable is stored in the portion of NVR allocated for error
-        /// records.
-        const HARDWARE_ERROR_RECORD = 0x08;
-
-        /// Deprecated.
-        const AUTHENTICATED_WRITE_ACCESS = 0x10;
-
-        /// Variable payload begins with an EFI_VARIABLE_AUTHENTICATION_2
-        /// structure.
-        const TIME_BASED_AUTHENTICATED_WRITE_ACCESS = 0x20;
-
-        /// This is never set in the attributes returned by
-        /// `get_variable`. When passed to `set_variable`, the variable payload
-        /// will be appended to the current value of the variable if supported
-        /// by the firmware.
-        const APPEND_WRITE = 0x40;
-
-        /// Variable payload begins with an EFI_VARIABLE_AUTHENTICATION_3
-        /// structure.
-        const ENHANCED_AUTHENTICATED_ACCESS = 0x80;
-    }
-}
-
-newtype_enum! {
-    /// Variable vendor GUID. This serves as a namespace for variables to
-    /// avoid naming conflicts between vendors. The UEFI specification
-    /// defines some special values, and vendors will define their own.
-    pub enum VariableVendor: Guid => {
-        /// Used to access global variables.
-        GLOBAL_VARIABLE = guid!("8be4df61-93ca-11d2-aa0d-00e098032b8c"),
-
-        /// Used to access EFI signature database variables.
-        IMAGE_SECURITY_DATABASE = guid!("d719b2cb-3d3a-4596-a3bc-dad00e67656f"),
-    }
-}
 
 /// Unique key for a variable.
 #[cfg(feature = "alloc")]
